move compartmentfitter to modelreduction directory
