+++ neat/__init__.py
@@ -54,6 +54,7 @@ from .tools.kernelextraction import FourrierTools
 
 from .channels.ionchannels import IonChannel
 
-from .tools.fittools.compartmentfitter import CompartmentFitter
+from .modelreduction.compartmentfitter import CompartmentFitter
+from .modelreduction.cachetrees import EquilibriumTree
 
 from .__version__ import __version__
